refactoring

+++ app/Http/Controllers/IssueController.php
@@ -10,37 +10,8 @@ use App\Http\Traits\ResponseTrait;
 class IssueController extends Controller
 {
     use ResponseTrait;
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function index()
-    {
-        //
-    }
-
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        //
-    }
-
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
-    public function store(Request $request)
-    {
-        //
-    }
-
+ 
+    
     /**
      * Display the specified resource.
      *
@@ -49,20 +20,25 @@ class IssueController extends Controller
      */
     public function show($id)
     {
-        //
-    }
+        try {
 
+            $issueExist = Issue::find($id);
+            if (!$issueExist) {
+                return $this->error_response("Not found", 404);
+            }
+            return $this->success_response($issueExist, 200);
+        } catch (Exception $e) {
+            return $this->error_response($e->getMessage(), 500);
+        }
+    }
+    
     /**
-     * Show the form for editing the specified resource.
+     * change status of the specified issue in storage
      *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
+     * @param  \Illuminate\Http\Request  $request
+     * @param  int $id
+     *  @return \Illuminate\Http\Response
      */
-    public function edit($id)
-    {
-        //
-    }
-
     public function change_status(Request $request, $id)
     {
         try {
@@ -77,7 +53,7 @@ class IssueController extends Controller
             }
             $issueExist["status"] = $request->status;
             if ($issueExist->save()) {
-                return $this->ok_response("status updated", 200);
+                return $this->success_response("status updated", 200);
             } else {
                 return $this->error_response("error in status update", 400);
             }
@@ -95,7 +71,33 @@ class IssueController extends Controller
      */
     public function update(Request $request, $id)
     {
-        //
+        try {
+            $this->validate($request, [
+                "name" => 'string',
+                "description" => 'string',
+                "approved" => 'boolean'
+            ]);
+            $issueExist = Issue::find($id);
+            if (auth()->user()->id == $issueExist->created_by) {
+
+
+
+                if (!$issueExist) {
+                    return $this->error_response("Not found", 404);
+                }
+                $issueExist->fill($request->all());
+                $issueExist["updated_by"] = auth()->user()->id;
+                if ($issueExist->save()) {
+                    return $this->success_response("issue updated", 200);
+                } else {
+                    return $this->error_response("error in status update", 400);
+                }
+            } else {
+                return $this->error_response("Forbidden", 403);
+            }
+        } catch (Exception $e) {
+            return $this->error_response($e->getMessage(), 500);
+        }
     }
 
     /**
@@ -108,15 +110,21 @@ class IssueController extends Controller
     {
         try {
             // if(auth()->user()->can()){
+
             $issueExist = Issue::find($id);
-            if (!$issueExist) {
-                return $this->error_response('Not found', 404);
-            }
-            if ($issueExist->delete()) {
-                return $this->ok_response("", 204);
+            if (auth()->user()->id == $issueExist->created_by) {
+                if (!$issueExist) {
+                    return $this->error_response('Not found', 404);
+                }
+                if ($issueExist->delete()) {
+                    return $this->success_response("", 204);
+                } else {
+                    return $this->error_response("Error in delete", 400);
+                }
             } else {
-                return $this->error_response("Error in delete", 400);
+                return $this->error_response("Forbidden", 403);
             }
+
             // }else{
 
             // }
